Trailing bits of prefix MUST be zero
This requires WG discussion more than some of the changes: it's a normative change to require the trailing bits to be zero. My guess is that all implementations do this anyway because they zero memory on allocation. If so, we should make this change. If we are unsure, I would propose downgrading to SHOULD and then upgrade it to MUST next time we touch the document.

As part of this checkin, also eliminate the redundant definition of the prefix format. We only need it once, it's better to reference it elsewhere.