Reset Comparison Section
I don't see why we need to drag a complete changelog back to the beginning of time along in this document. Apart from needlessly inflating the page count, it requires a bunch of references that are otherwise not required. Readers wanting to know the older history can always go look at RFC 4271.